requirements.txt
haystack-ai
mcp-haystack
amazon-bedrock-haystack
langfuse-haystack
questionary